handle_bot_message/requirements.txt
boto3~=1.21.22
pytz~=2021.3
botocore~=1.24.22
gspread~=5.2.0
python-telegram-bot
flask
boto3-stubs